requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.5
selenium>=4.15.2
httpx>=0.25.2
aiohttp>=3.9.1

# Data validation
pydantic>=2.5.0
email-validator>=2.1.0
phonenumbers>=8.13.0

# DNS and networking
dnspython>=2.4.2
python-whois>=0.8.0
netaddr>=0.9.0

# CLI and formatting
rich>=13.7.0
click>=8.1.7
tabulate>=0.9.0
colorama>=0.4.6

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Web scraping helpers
cloudscraper>=1.2.71
fake-useragent>=1.4.0
httplib2>=0.22.0

# Database
sqlalchemy>=2.0.23

# Utilities
pytz>=2023.3
tenacity>=8.2.3

# OSINT specific (optionnel - à installer manuellement si disponible)
# holehe>=1.61.0
# socialscan>=1.4.2
# maigret>=0.4.4
# pwnedpasswords>=2.0.0
# h8mail>=2.5.3